+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated hour after event label text in Event2SetupSurveyPanel [#4784](https://github.com/rokwire/illinois-app/issues/4784).
 - Show additional attendance takers control in create event / setup attendance [#4777](https://github.com/rokwire/illinois-app/issues/4777).
 - Updated Event Type and Location layout in Create/Edit Event panel [#4775](https://github.com/rokwire/illinois-app/issues/4775).
+### Added
+- Implemented downloading survey responses in the web app [#4799](https://github.com/rokwire/illinois-app/issues/4799).
 
 ## [6.1.61] - 2025-02-17
 ### Changed